+++ src/GRA.Data/Profile/MappingProfile.cs
@@ -59,6 +59,8 @@ namespace GRA.Data.Profile
             CreateMap<Model.Event, Domain.Model.Event>()
                 .ForMember(dest => dest.Challenge, opt => opt.Ignore())
                 .ForMember(dest => dest.ChallengeGroup, opt => opt.Ignore());
+            CreateMap<Model.FeaturedChallengeGroup, Domain.Model.FeaturedChallengeGroup>();
+            CreateMap<Model.FeaturedChallengeGroupText, Domain.Model.FeaturedChallengeGroupText>();
             CreateMap<Model.GroupInfo, Domain.Model.GroupInfo>();
             CreateMap<Model.GroupType, Domain.Model.GroupType>();
             CreateMap<Model.Job, Domain.Model.Job>();
@@ -162,6 +164,8 @@ namespace GRA.Data.Profile
             CreateMap<Domain.Model.EmailReminder, Model.EmailReminder>();
             CreateMap<Domain.Model.EmailSubscriptionAuditLog, Model.EmailSubscriptionAuditLog>();
             CreateMap<Domain.Model.Event, Model.Event>();
+            CreateMap<Domain.Model.FeaturedChallengeGroup, Model.FeaturedChallengeGroup>();
+            CreateMap<Domain.Model.FeaturedChallengeGroupText, Model.FeaturedChallengeGroupText>();
             CreateMap<Domain.Model.GroupInfo, Model.GroupInfo>();
             CreateMap<Domain.Model.GroupType, Model.GroupType>();
             CreateMap<Domain.Model.Job, Model.Job>();