+++ src/GRA.Domain.Model/DisplayChallengeGroup.cs
@@ -1,0 +1,12 @@
+﻿namespace GRA.Domain.Model
+{
+    public class DisplayChallengeGroup
+    {
+        public string AltText { get; set; }
+        public string ImageFile { get; set; }
+        public string ImagePath { get; set; }
+        public string Name { get; set; }
+        public int SortOrder { get; set; }
+        public string Stub { get; set; }
+    }
+}